Fixes #689: Arrange services in a separate folder
