+++ src/app/app.component.spec.ts
@@ -29,7 +29,7 @@ import { DropdownComponent } from './dropdown/dropdown.component';
 import {IntelligenceComponent} from "./intelligence/intelligence.component";
 import {SpeechtotextComponent} from "./speechtotext/speechtotext.component";
 import {AutoCorrectComponent} from "./auto-correct/auto-correct.component";
-import {SpeechService} from "./speech.service";
+import {SpeechService} from "./services/speech.service";
 
 describe('AppComponent', () => {
   beforeEach(() => {
