+++ src/app/app.module.ts
@@ -7,7 +7,7 @@ import { NavbarComponent } from './navbar/navbar.component';
 import { IndexComponent } from './index/index.component';
 import { ResultsComponent } from './results/results.component';
 import { Routes, RouterModule } from '@angular/router';
-import { SearchService } from './search.service';
+import { SearchService } from './services/search.service';
 import { NotFoundComponent } from './not-found/not-found.component';
 import { CommonModule } from '@angular/common';
 import { AdvancedsearchComponent } from './advancedsearch/advancedsearch.component';
@@ -24,24 +24,24 @@ import { EffectsModule } from '@ngrx/effects';
 import { ApiSearchEffects } from './effects/search-effects';
 import { NewadvancedsearchComponent } from './newadvancedsearch/newadvancedsearch.component';
 import { InfoboxComponent } from './infobox/infobox.component';
-import { KnowledgeapiService } from './knowledgeapi.service';
+import { KnowledgeapiService } from './services/knowledgeapi.service';
 import { RelatedSearchComponent } from './related-search/related-search.component';
-import { AutocompleteService } from "./autocomplete.service";
+import { AutocompleteService } from "./services/autocomplete.service";
 import { AutoCompleteComponent } from './auto-complete/auto-complete.component';
 import { ThemeComponent } from './theme/theme.component';
-import { ThemeService } from './theme.service';
+import { ThemeService } from './services/theme.service';
 import { CrawlstartComponent } from './crawlstart/crawlstart.component';
-import { CrawlstartService } from "./crawlstart.service";
+import { CrawlstartService } from "./services/crawlstart.service";
 import { SearchsettingsComponent } from './searchsettings/searchsettings.component';
 import { KnowledgeEffects } from "./effects/knowledge";
-import { SpeechService } from './speech.service';
+import { SpeechService } from './services/speech.service';
 import { DropdownComponent } from './dropdown/dropdown.component';
 import { IntelligenceComponent } from './intelligence/intelligence.component';
-import {IntelligenceService} from "./intelligence.service";
+import {IntelligenceService} from "./services/intelligence.service";
 import { SpeechtotextComponent } from './speechtotext/speechtotext.component';
 import { AutoCorrectComponent } from './auto-correct/auto-correct.component';
-import {AutocorrectService} from "./autocorrect.service";
-import { SpeechSynthesisService } from './speech-synthesis.service';
+import {AutocorrectService} from "./services/autocorrect.service";
+import { SpeechSynthesisService } from './services/speech-synthesis.service';
 
 const appRoutes: Routes = [
   {path: 'search', component: ResultsComponent},
