+++ src/app/auto-complete/auto-complete.component.ts
@@ -1,5 +1,5 @@
 import {Component, OnInit, ChangeDetectorRef, Output, EventEmitter} from '@angular/core';
-import {AutocompleteService} from "../autocomplete.service";
+import {AutocompleteService} from "../services/autocomplete.service";
 import {Router, ActivatedRoute} from "@angular/router";
 import {Store} from "@ngrx/store";
 import * as fromRoot from '../reducers';
