+++ src/app/infobox/infobox.component.spec.ts
@@ -3,8 +3,8 @@ import { HttpModule, JsonpModule } from '@angular/http';
 import { InfoboxComponent } from './infobox.component';
 import { RouterTestingModule } from "@angular/router/testing";
 
-import { KnowledgeapiService } from "../knowledgeapi.service";
-import { SpeechSynthesisService } from "../speech-synthesis.service";
+import { KnowledgeapiService } from "../services/knowledgeapi.service";
+import { SpeechSynthesisService } from "../services/speech-synthesis.service";
 import { reducer } from "../reducers/index";
 import { StoreModule } from "@ngrx/store";
 import { MockKnowledgeApi } from "../shared/mocks/knowledge.mock";
