+++ src/app/intelligence/intelligence.component.spec.ts
@@ -17,7 +17,7 @@ import {BrowserModule} from "@angular/platform-browser";
 import {CommonModule} from "@angular/common";
 import {FormsModule} from "@angular/forms";
 import {HttpModule, JsonpModule} from "@angular/http";
-import {KnowledgeapiService} from "../knowledgeapi.service";
+import {KnowledgeapiService} from "../services/knowledgeapi.service";
 import {reducer} from "../reducers/index";
 import {StoreModule} from "@ngrx/store";
 import {StoreDevtoolsModule} from "@ngrx/store-devtools";
@@ -26,7 +26,7 @@ import {AutoCompleteComponent} from "../auto-complete/auto-complete.component";
 import { ThemeComponent } from '../theme/theme.component';
 import { DropdownComponent } from '../dropdown/dropdown.component';
 import {IntelligenceComponent} from "../intelligence/intelligence.component";
-import {IntelligenceService} from "../intelligence.service";
+import {IntelligenceService} from "../services/intelligence.service";
 import {InfoboxComponent} from "../infobox/infobox.component";
 import {AutoCorrectComponent} from "../auto-correct/auto-correct.component";
