+++ src/app/intelligence/intelligence.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import {Observable} from "rxjs";
 import {Store} from "@ngrx/store";
 import * as fromRoot from '../reducers';
-import {IntelligenceService} from "../intelligence.service";
+import {IntelligenceService} from "../services/intelligence.service";
 
 @Component({
   selector: 'app-intelligence',
