+++ src/app/not-found/not-found.component.spec.ts
@@ -24,9 +24,9 @@ import { ContactComponent } from '../contact/contact.component';
 import {InfoboxComponent} from "../infobox/infobox.component";
 import {RelatedSearchComponent} from "../related-search/related-search.component";
 import {AutoCompleteComponent} from "../auto-complete/auto-complete.component";
-import {AutocompleteService} from "../autocomplete.service";
+import {AutocompleteService} from "../services/autocomplete.service";
 import { ThemeComponent } from '../theme/theme.component';
-import { SpeechService } from '../speech.service';
+import { SpeechService } from '../services/speech.service';
 import { DropdownComponent } from '../dropdown/dropdown.component';
 import {IntelligenceComponent} from "../intelligence/intelligence.component";
 import {SpeechtotextComponent} from "../speechtotext/speechtotext.component";
