+++ src/app/related-search/related-search.component.ts
@@ -2,7 +2,7 @@ import {Component, OnInit, ChangeDetectorRef} from '@angular/core';
 import {Router, ActivatedRoute} from '@angular/router';
 import {Store} from '@ngrx/store';
 import * as fromRoot from '../reducers';
-import {KnowledgeapiService} from '../knowledgeapi.service';
+import {KnowledgeapiService} from '../services/knowledgeapi.service';
 import {Observable} from "rxjs";
 
 @Component({
