+++ src/app/results/results.component.spec.ts
@@ -17,26 +17,26 @@ import { IndexComponent } from '../index/index.component';
 import { NotFoundComponent } from '../not-found/not-found.component';
 import { AdvancedsearchComponent } from '../advancedsearch/advancedsearch.component';
 import { SearchBarComponent } from '../search-bar/search-bar.component';
-import { SearchService } from '../search.service';
+import { SearchService } from '../services/search.service';
 import { FooterNavbarComponent } from '../footer-navbar/footer-navbar.component';
 import { AboutComponent } from '../about/about.component';
 import { ModalComponent, Ng2Bs3ModalModule } from 'ng2-bs3-modal/ng2-bs3-modal';
 import { ContactComponent } from '../contact/contact.component';
 import {InfoboxComponent} from "../infobox/infobox.component";
-import {KnowledgeapiService} from "../knowledgeapi.service";
+import {KnowledgeapiService} from "../services/knowledgeapi.service";
 import {RelatedSearchComponent} from "../related-search/related-search.component";
 import {AutoCompleteComponent} from "../auto-complete/auto-complete.component";
-import {AutocompleteService} from "../autocomplete.service";
+import {AutocompleteService} from "../services/autocomplete.service";
 import { ThemeComponent } from '../theme/theme.component';
-import { ThemeService } from '../theme.service';
-import { SpeechService } from '../speech.service';
+import { ThemeService } from '../services/theme.service';
+import { SpeechService } from '../services/speech.service';
 import { DropdownComponent } from '../dropdown/dropdown.component';
 import {IntelligenceComponent} from "../intelligence/intelligence.component";
-import {IntelligenceService} from "../intelligence.service";
+import {IntelligenceService} from "../services/intelligence.service";
 import {SpeechtotextComponent} from "../speechtotext/speechtotext.component";
 import {AutoCorrectComponent} from "../auto-correct/auto-correct.component";
-import {AutocorrectService} from "../autocorrect.service";
-import { SpeechSynthesisService } from "../speech-synthesis.service";
+import {AutocorrectService} from "../services/autocorrect.service";
+import { SpeechSynthesisService } from "../services/speech-synthesis.service";
 
 describe('ResultsComponent', () => {
   let component: ResultsComponent;
