+++ src/app/services/autocomplete.service.spec.ts
@@ -1,12 +1,14 @@
+/* tslint:disable:no-unused-variable */
+
 import { TestBed, inject } from '@angular/core/testing';
+import { AutocompleteService } from './autocomplete.service';
 import { RouterTestingModule } from '@angular/router/testing';
 import {HttpModule, JsonpModule} from "@angular/http";
 import {StoreModule} from "@ngrx/store";
 import {StoreDevtoolsModule} from "@ngrx/store-devtools";
-import {reducer} from "./reducers/index";
-import {AutocorrectService} from "./autocorrect.service";
+import {reducer} from "../reducers/index";
 
-describe('AutocorrectService', () => {
+describe('AutocompleteService', () => {
   beforeEach(() => {
     TestBed.configureTestingModule({
       imports: [
@@ -15,11 +17,11 @@ describe('AutocorrectService', () => {
         StoreModule.provideStore(reducer),
         StoreDevtoolsModule.instrumentOnlyWithExtension(),
       ],
-      providers: [AutocorrectService]
+      providers: [AutocompleteService]
     });
   });
 
-  it('should be created', inject([AutocorrectService], (service: AutocorrectService) => {
+  it('should ...', inject([AutocompleteService], (service: AutocompleteService) => {
     expect(service).toBeTruthy();
   }));
 });
