+++ src/app/services/search.service.ts
@@ -3,8 +3,8 @@ import {Http, URLSearchParams, Jsonp, Response} from '@angular/http';
 import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/toPromise';
 import {Store} from '@ngrx/store';
-import * as fromRoot from './reducers';
-import * as search from './actions/search';
+import * as fromRoot from '../reducers';
+import * as search from '../actions/search';
 import {Observable} from 'rxjs';
 
 @Injectable()
