+++ src/app/services/speech.service.ts
