+++ src/app/speechtotext/speechtotext.component.spec.ts
@@ -1,7 +1,7 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { SpeechtotextComponent } from './speechtotext.component';
-import {SpeechService} from "../speech.service";
+import {SpeechService} from "../services/speech.service";
 import {StoreModule} from "@ngrx/store";
 import {reducer} from "../reducers/index";
 import {RouterTestingModule} from "@angular/router/testing";
